module-6-7-8-9/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14
psycopg2-binary>=2.9
django-environ>=0.11
drf-spectacular>=0.27.0
gunicorn>=21.2
